// Serie.d.ts
export declare class Serie {
    id: number;
    name: string;
    channel: string;
    seasons: number;
    description: string;
    image: string;
    url: string;
    constructor(id: number, name: string, channel: string, seasons: number, description: string, image: string, url: string);
}
//# sourceMappingURL=Serie.d.ts.map